# Fireblocks.swagger.yml
...
openapi: 3.0.0
info:
  title: Fireblocks API
  version: "1.5.0"
servers:
  - url: 'https://api.fireblocks.io/v1'
paths:
...
  '/vault/assets':
    get:
      summary: Gets the assets amount summary for all \ filtered accounts.
      parameters:
        - in: query
          name: accountNamePrefix
          required: false
          schema:
            type: string
        - in: query
          name: accountNameSuffix
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Amount by asset
          content:
            '*/*':
              schema:
                  type: array
                  items:
                    $ref: '#/components/schemas/VaultAsset'
        default:
          $ref: '#/components/responses/Error'
  '/vault/assets/{assetId}':
    get:
      summary: Gets vault balance summary by asset.
      parameters:
        - in: path
          name: assetId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Vault amount by asset
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/VaultAsset'
        default:
          $ref: '#/components/responses/Error'
  #  INTERNAL WALLETS
...
  '/transactions/{txId}':
    get:
      summary: Returns a transaction by ID
      parameters:
        - in: path
          name: txId
          required: true
          description: The ID of the transaction to return
          schema:
            type: string
            minimum: 1
      responses:
        '200':
          description: An Transaction object
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/TransactionResponse'
        default:
          $ref: '#/components/responses/Error'
...
  '/transactions/{txId}/cancel':
    post:
      summary: Cancels a transaction by ID
      parameters:
        - in: path
          name: txId
          required: true
          description: The ID of the transaction to cancel
          schema:
            type: string
            minimum: 1
      responses:
        '200':
          description: An Transaction object
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CancelTransactionResponse'
        default:
          $ref: '#/components/responses/Error'
...
  '/supported_assets':
    get:
      summary: Returns all asset types supported by Fireblocks
      responses:
        '200':
          description: A Transaction object
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AssetTypeResponse'
        default:
          $ref: '#/components/responses/Error'
  # NETWORK CONNECTIONS
...
security:
  - bearerTokenAuth: [ ]
  - ApiKeyAuth: [ ]
components:
  responses:
    Error:
      description: Error Response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerTokenAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    AllocatedBalance:
      type: object
      properties:
        allocationId:
          type: string
        thirdPartyAccountId:
          type: string
        affiliation:
          type: string
        virtualType:
          type: string
        total:
          type: string
        available:
          type: string
        pending:
          type: string
        frozen:
          type: string
        locked:
          type: string
        staked:
          type: string
    VaultAsset:
      type: object
      properties:
        id:
          type: string
        total:
          description: The total wallet balance. In EOS this value includes the network balance, self staking and pending refund. For all other coins it is the balance as it appears on the blockchain.
          type: string
        balance:
          deprecated: true
          description: Deprecated - replaced by "total"
          type: string
        available:
          description: Funds available for transfer. Equals the blockchain balance minus any locked amounts
          type: string
        pending:
          description: The cumulative balance of all transactions pending to be cleared
          type: string
        frozen:
          description: The cumulative frozen balance
          type: string
        lockedAmount:
          description: Funds in outgoing transactions that are not yet published to the network
          type: string
        staked:
          description: Staked balance
          type: string
        totalStakedCPU:
          type: string
        totalStakedNetwork:
          type: string
        selfStakedCPU:
          type: string
        selfStakedNetwork:
          type: string
        pendingRefundCPU:
          type: string
        pendingRefundNetwork:
          type: string
        blockHeight:
          type: string
        blockHash:
          type: string
          nullable: true
        allocatedBalances:
          type: array
          items:
            $ref: '#/components/schemas/AllocatedBalance'
...
    TransferPeerPath:
      type: object
      properties:
        type:
          type: string
        id:
          type: string
          nullable: true
      required:
        - type
...
    CancelTransactionResponse:
      type: object
      properties:
        success:
          type: boolean
...
    AmlScreeningResult:
      type: object
      properties:
        provider:
          type: string
        payload:
          type: object
...
    BlockInfo:
      type: object
      properties:
        blockHeight:
          type: string
          nullable: true
        blockHash:
          type: string
          nullable: true
    AuthorizationInfo:
      type: object
      properties:
        allowOperatorAsAuthorizer:
          type: boolean
        logic:
          type: string
          enum:
            - AND
            - OR
        groups:
          type: array
          items:
            $ref: '#/components/schemas/AuthorizationGroups'
    AuthorizationGroups:
      type: object
      properties:
        th:
          type: number
          format: decimal
        users:
          type: object
          additionalProperties:
            type: string
            enum:
              - PENDING_AUTHORIZATION
              - APPROVED
              - REJECTED
              - NA
    AmountInfo:
      type: object
      properties:
        amount:
          type: string
          nullable: true
        requestedAmount:
          type: string
          nullable: true
        netAmount:
          type: string
          nullable: true
        amountUSD:
          type: string
          nullable: true
    TransferPeerPathResponse:
      allOf:
        - $ref: '#/components/schemas/TransferPeerPath'
        - type: object
          properties:
            name:
              type: string
            subType:
              type: string
            virtualType:
              type: string
              enum:
                - OFF_EXCHANGE
                - DEFAULT
                - UNKNOWN
            virtualId:
              type: string
    # This object is not in the original schema. This was introduced to allow Stablehouse to poll
    # on the transactions endpoint while only deserializing the ID. This allows us to safely poll
    # for new transactions and not worry about serialization errors when fetching the transaction list.
...
    TransactionResponse:
      type: object
      properties:
        id:
          type: string
        assetId:
          type: string
          nullable: true
        source:
          $ref: '#/components/schemas/TransferPeerPathResponse'
          nullable: true
        destination:
          $ref: '#/components/schemas/TransferPeerPathResponse'
          nullable: true
        requestedAmount:
          description: The amount requested by the user
          type: string
          nullable: true
        amount:
          description: If the transfer is a withdrawal from an exchange, the actual amount that was requested to be transferred. Otherwise, the requested amount
          type: string
          nullable: true
        netAmount:
          description: The net amount of the transaction, after fee deduction
          type: string
          nullable: true
        amountUSD:
          description: The USD value of the requested amount
          type: string
          nullable: true
        serviceFee:
          description: The total fee deducted by the exchange from the actual requested amount (serviceFee = amount - netAmount)
          type: string
          nullable: true
        networkFee:
          description: The fee paid to the network
          type: string
          nullable: true
        createdAt:
          type: number
          format: decimal
          description: Unix timestamp
          nullable: true
        lastUpdated:
          type: number
          format: decimal
          description: Unix timestamp
        status:
          type: string
        txHash:
          type: string
        tag:
          type: string
        subStatus:
          $ref: '#/components/schemas/TransactionSubStatus'
          nullable: true
        destinationAddress:
          type: string
        sourceAddress:
          type: string
        destinationAddressDescription:
          type: string
        destinationTag:
          type: string
        signedBy:
          type: array
          items:
            type: string
        createdBy:
          type: string
        rejectedBy:
          type: string
        addressType:
          type: string
        note:
          type: string
        exchangeTxId:
          type: string
        feeCurrency:
          type: string
        operation:
          $ref: '#/components/schemas/TransactionOperation'
        networkRecords:
          type: array
          items:
            $ref: '#/components/schemas/NetworkRecord'
        amlScreeningResult:
          $ref: '#/components/schemas/AmlScreeningResult'
        customerRefId:
          type: string
        numOfConfirmations:
          type: number
          format: decimal
          nullable: true
        signedMessages:
          type: array
          items:
            $ref: '#/components/schemas/SignedMessage'
        extraParameters:
          type: object
          properties: {}
        externalTxId:
          type: string
        replacedTxHash:
          type: string
        destinations:
          type: array
          items:
            $ref: '#/components/schemas/TransactionResponseDestination'
        blockInfo:
          $ref: '#/components/schemas/BlockInfo'
        authorizationInfo:
          $ref: '#/components/schemas/AuthorizationInfo'
        amountInfo:
          $ref: '#/components/schemas/AmountInfo'
        index:
          type: number
          format: decimal
          nullable: true
        blockchainIndex:
          type: string
    TransactionResponseDestination:
      type: object
      properties:
        amount:
          type: string
        amountUSD:
          type: string
        amlScreeningResult:
          nullable: true
          $ref: '#/components/schemas/AmlScreeningResult'
        destination:
          $ref: '#/components/schemas/TransferPeerPathResponse'
        authorizationInfo:
          nullable: true
          $ref: '#/components/schemas/AuthorizationInfo'
    NetworkRecord:
      type: object
      properties:
        source:
          $ref: '#/components/schemas/TransferPeerPathResponse'
        destination:
          $ref: '#/components/schemas/TransferPeerPathResponse'
        txHash:
          type: string
        networkFee:
          type: string
        assetId:
          type: string
        netAmount:
          description: The net amount of the transaction, after fee deduction
          type: string
        isDropped:
          type: boolean
        type:
          type: string
        destinationAddress:
          type: string
        sourceAddress:
          type: string
          nullable: true
        amountUSD:
          type: string
          nullable: true
        index:
          type: number
          format: decimal
          nullable: true
    AssetTypeResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
        contractAddress:
          type: string
        issuerAddress:
          type: string
        nativeAsset:
          type: string
        decimals:
          type: number
          format: decimal
      required:
        - id
        - name
        - type
...
    TransactionOperation:
      type: string
      description: >
        - TRANSFER
        - MINT
        - BURN
        - SUPPLY_TO_COMPOUND
        - REDEEM_FROM_COMPOUND
        - RAW
        - CONTRACT_CALL
        - INTERNAL_LEDGER_TRANSFER
        - TYPED_MESSAGE
      default: TRANSFER
    Error:
      type: object
      properties:
        message:
          type: string
        code:
          type: number
          format: decimal
    TransactionSubStatus:
      type: string
      description: >
        - `INSUFFICIENT_FUNDS` - Not enough funds to fulfill the withdraw request

        - `AMOUNT_TOO_SMALL` - Attempt to withdraw an amount below the allowed minimum

        - `UNSUPPORTED_ASSET` - Asset is not supported

        - `UNAUTHORISED__MISSING_PERMISSION` - Third party (e.g. exchange) API missing permission

        - `INVALID_SIGNATURE` - Invalid transaction signature

        - `API_INVALID_SIGNATURE` - Third party (e.g. exchange) API call invalid signature

        - `UNAUTHORISED__MISSING_CREDENTIALS` - Missing third party (e.g. exchange) credentials

        - `UNAUTHORISED__USER` - Attempt to initiate or approve a transaction by an unauthorised user

        - `UNAUTHORISED__DEVICE` - Unauthorised user's device

        - `INVALID_UNMANAGED_WALLET` - Unmanaged wallet is disabled or does not exist

        - `INVALID_EXCHANGE_ACCOUNT` - Exchange account is disabled or does not exist

        - `INSUFFICIENT_FUNDS_FOR_FEE` - Not enough balance to fund the requested transaction

        - `INVALID_ADDRESS` - Unsupported address format

        - `WITHDRAW_LIMIT` - Transaction exceeds the exchange's withdraw limit

        - `API_CALL_LIMIT` - Exceeded third party (e.g. exchange) API call limit

        - `ADDRESS_NOT_WHITELISTED` - Attempt to withdraw from an exchange to a non whitelisted address

        - `TIMEOUT` - Request timeout

        - `CONNECTIVITY_ERROR` - Network error

        - `THIRD_PARTY_INTERNAL_ERROR` - Received an internal error response from a third party service

        - `CANCELLED_EXTERNALLY` - Transaction was canceled by a third party service

        - `INVALID_THIRD_PARTY_RESPONSE` - Unrecognized third party response

        - `VAULT_WALLET_NOT_READY` - Vault wallet is not ready

        - `MISSING_DEPOSIT_ADDRESS` - Could not retrieve a deposit address from the exchange

        - `INTERNAL_ERROR` - Internal error while processing the transaction

        - `UNKNOWN_ERROR` - Unexpected error

        - `AUTHORIZER_NOT_FOUND` - No authorizer found to approve the operation or the only authorizer found is the initiator

        - `INSUFFICIENT_RESERVED_FUNDING` - Some assets require a minimum of reserved funds being kept on the account

        - `MANUAL_DEPOSIT_ADDRESS_REQUIRED` - Error while retrieving a deposit address from an exchange. Please generate a deposit address for your exchange account

        - `INVALID_FEE` - Transaction fee is not in the allowed range

        - `ERROR_UNSUPPORTED_TRANSACTION_TYPE` - Attempt to execute an unsupported transaction type

        - `UNSUPPORTED_OPERATION` - Unsupported operation

        - `3RD_PARTY_PROCESSING` - The transaction is pending approval by the 3rd party service (e.g exchange)

        - `PENDING_BLOCKCHAIN_CONFIRMATIONS` - Pending Blockchain confirmations

        - `3RD_PARTY_CONFIRMING` - Pending confirmation on the exchange

        - `CONFIRMED` - Confirmed on the blockchain

        - `3RD_PARTY_COMPLETED` - Completed on the 3rd party service (e.g exchange)

        - `PENDING_BLOCKCHAIN_CONFIRMATIONSREJECTED_BY_USER` - The transaction was rejected by one of the signers

        - `CANCELLED_BY_USER` - The transaction was canceled via the Console or the API

        - `3RD_PARTY_CANCELLED` - Cancelled on the exchange

        - `3RD_PARTY_REJECTED` - Rejected or not approved in time by user

        - `AML_SCREENING_REJECTED` - Rejected on AML Screening

        - `BLOCKED_BY_POLICY` - Transaction is blocked due to a policy rule

        - `FAILED_AML_SCREENING` - AML Screening failed

        - `PARTIALLY_FAILED` - Only for Aggregated transactions. One or more of the associated transaction records failed

        - `3RD_PARTY_FAILED` - Transaction failed at the exchange

        - `ZERO_BALANCE_IN_PERMANENT_ADDRESS` - Not enough BTC balance to fund the requested transaction in legacy address

        - 'DROPPED_BY_BLOCKCHAIN' - The transaction was replaced by another transaction with higher fee

        - 'INVALID_FEE_PARAMS' - Fee parameters are inconsistent or unknown

        - 'MISSING_TAG_OR_MEMO' - A tag or memo is required to send funds to a third party address, including all exchanges

        - 'SIGNING_ERROR' - The transaction signing failed, resubmit the transaction to sign again

        - 'GAS_LIMIT_TOO_LOW' - The transaction was rejected because the gas limit was set too low

        -  'MAX_FEE_EXCEEDED' - Gas price is currently above selected max fee

        - 'ACTUAL_FEE_TOO_HIGH' - Chosen fee is below current price

        - 'INVALID_CONTRACT_CALL_DATA' - Transaction data was not encoded properly

        - 'INVALID_NONCE_TOO_LOW' - Illegal nonce

        - 'INVALID_NONCE_TOO_HIGH' - Illegal nonce

        - 'INVALID_NONCE_FOR_RBF' - No matching nonce

        - 'FAIL_ON_LOW_FEE' - Current blockchain fee is higher than selected

        - 'TOO_LONG_MEMPOOL_CHAIN' - Too many unconfirmed transactions from this address

        - 'TX_OUTDATED' - Nonce already used

        - 'INCOMPLETE_USER_SETUP' - MPC setup was not completed

        - 'SIGNER_NOT_FOUND' - Signer not found

        - 'INVALID_TAG_OR_MEMO' - Invalid Tag or Memo

        - 'ZERO_BALANCE_IN_PERMANENT_ADDRESS' - Not enough BTC on legacy permanent address

        - 'NEED_MORE_TO_CREATE_DESTINATION' - Insufficient funds for creating destination account

        - 'NON_EXISTING_ACCOUNT_NAME' - Account does not exist

        - 'ENV_UNSUPPORTED_ASSET' - Asset is not supported under this workspace settings
...
    SignedMessage:
      type: object
      properties:
        content:
          type: string
        algorithm:
          type: string
          enum:
            - MPC_ECDSA_SECP256K1
            - MPC_ECDSA_SECP256R1
            - MPC_EDDSA_ED25519
        derivationPath:
          type: array
          items:
            type: number
            format: decimal
        signature:
          type: object
          properties:
            fullSig:
              type: string
            r:
              type: string
            s:
              type: string
            v:
              type: number
              format: decimal
        publicKey:
          type: string
...